Refactored show action and added handling error when bad id in URL

+++ app/controllers/readers_controller.rb
@@ -2,11 +2,14 @@ class ReadersController < ApplicationController
   before_action :authenticate_reader!
 
   def show
-    @reader = Reader.new
-    if session[:transaction_id]
-      t = Transaction.find(session[:transaction_id])
-      @book = Book.find(t.book_id)
-      session.delete(:transaction_id)
+    @reader = Reader.find_by(id: params[:id])
+    if @reader.present?
+      if session[:transaction_id]
+        @book = Transaction.find(session[:transaction_id]).book
+        session.delete(:transaction_id)
+      end
+    else
+      redirect_to root_path, notice: "No hemos encontrado a ningún usuario con ese ID"
     end
   end
 
@@ -14,14 +17,14 @@ class ReadersController < ApplicationController
     if reader_params != nil
       respond_to do |format|
         if current_reader.update(reader_params)
-          format.html { redirect_to reader_path , notice: "Your info was succesfully updated" }
+          format.html { redirect_to reader_path , notice: "Tomamos nota. La información ha sido actualizada" }
         else
           flash[:error] = "Hmmm...algo no ha ido como esperábamos, asegúrate de rellenar los campos obligatorios!"
           format.html { redirect_to reader_path }
         end
       end
     else
-      redirect_to reader_path, notice: "Nada que actualizar, sigues siendo el mismo :) "
+      redirect_to reader_path, notice: "Nada que actualizar, sigues siendo el mismo"
     end
   end
 
@@ -32,7 +35,7 @@ class ReadersController < ApplicationController
       @books_sold = Transaction.get_seller_books(@reader)
       @books_acquired = Transaction.get_buyer_books(@reader)
     else
-      format.json { redirect_to reader_path, notice: "No user was found with that ID" }
+      redirect_to root_path, notice: "No user was found with that ID"
     end
   end
